pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-protocol-odata-py-lib"
version = "0.1.0-alpha.1"
description = "A Python package from the DS library collection"
dependencies = [
    "ds-resource-plugin-py-lib>=0.1.0-rc.3,<1.0.0",
    "ds-protocol-http-py-lib>=0.1.0b4,<1.0.0",
    "requests>=2.32.0,<3.0.0",
]
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE-APACHE"]
keywords = ["ds", "python"]
authors = [
    {name = "Grasp Labs AS", email = "hello@grasplabs.com"}
]
maintainers = [
    {name = "Grasp Labs AS", email = "hello@grasplabs.com"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
readme = "PyPI.md"

[project.urls]
Documentation = "https://grasp-labs.github.io/ds-protocol-odata-py-lib/"
Repository = "https://github.com/grasp-labs/ds-protocol-odata-py-lib/"
Issues = "https://github.com/grasp-labs/ds-protocol-odata-py-lib/issues/"
Changelog = "https://github.com/grasp-labs/ds-protocol-odata-py-lib/releases"

[project.entry-points."ds.protocols"]
odata = "ds_protocol_odata_py_lib"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pandas-stubs>=2.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "pre-commit>=3.6.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "types-requests>=2.32.0,<3.0.0",
    "sphinx>=7.1.0",
    "sphinx-autoapi>=3.0.0",
    "sphinx-material>=0.0.35",
    "bandit>=1.7.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
]
docs = [
    "sphinx>=7.1.0",
    "sphinx-autoapi>=3.0.0",
    "sphinx-material>=0.0.35",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
ds_protocol_odata_py_lib = ["resource.yaml", "py.typed"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
